test: move files
